Changed to get double

+++ backend_api/GnomadAPI/Repositories/Implementations/PinRepository.cs
@@ -59,8 +59,8 @@ namespace TravelCompanionAPI.Data
                         pin = new Pin();
                         pin.Id = reader.GetInt32(0);
                         pin.UserId = reader.GetInt32(1);
-                        pin.Longitude = reader.GetInt32(2);
-                        pin.Latitude = reader.GetInt32(3);
+                        pin.Longitude = reader.GetDouble(2);
+                        pin.Latitude = reader.GetDouble(3);
                         pin.Title = reader.GetString(4);
                         pin.Street = reader.GetString(5);
                     }
